requirements-test.txt
# test requirements
pytest
pytest-cov
pylama
pylama_pylint
tox
mock

-e .

# Install non-recommended keyring backends (for travis)
keyrings.alt